requirements.txt
streamlit==1.28.1
watchdog==6.0.0
PyMuPDF
python-dotenv
google-genai 
boto3==1.38.36

doclayout-yolo==0.0.4
huggingface_hub
pdf2image==1.16.3

